+++ test/unit/typesystem_isSubtype.cpp
@@ -10,9 +10,9 @@ using namespace xqp;
 
 int typesystem_isSubtype(int argc, char* argv[]) {
   Store& store = Store::getInstance();
-  Item_t lInteger = store.getItemFactory().createInteger(1);
+  Item_t lInteger = store.getItemFactory().createInteger((int32_t)1);
   Item_t lInt = store.getItemFactory().createInt(1);
-  Item_t lDecimal = store.getItemFactory().createDecimal(1);
+  Item_t lDecimal = store.getItemFactory().createDecimal((int32_t)1);
   
   TypeSystem::xqtref_t lIntegerType = GENV_TYPESYSTEM.create_type(lInteger->getType(), TypeSystem::QUANT_ONE);
   TypeSystem::xqtref_t lIntType = GENV_TYPESYSTEM.create_type(lInt->getType(), TypeSystem::QUANT_ONE);
